chore: update Go version and linter in GitHub Actions workflow
The Go version and the linter version in the GitHub Actions workflow have been updated. The Go version was changed from 'stable' to '1.21' to ensure compatibility with recent codebase changes. The linter version was also updated from v1.53 to v1.54 to take advantage of new linting features and fix detected false positives.